+++ dpnp/linalg/dpnp_iface_linalg.py
@@ -195,6 +195,14 @@ def cond(x, p=None):
     --------
     :obj:`dpnp.linalg.norm` : Matrix or vector norm.
 
+    Notes
+    -----
+    This function will raise :class:`dpnp.linalg.LinAlgError` on singular input
+    when using any of the norm: ``1``, ``-1``, ``inf``, ``-inf``, or ``'fro'``.
+    In contrast, :obj:`numpy.linalg.cond` will fill the result array with
+    ``inf`` values for each 2D batch in the input array that is singular
+    when using these norms.
+
     Examples
     --------
     >>> import dpnp as np